splits fingerprinter and tlsdecoder
